newone

+++ contracts/ReceiverContract.sol
@@ -5,11 +5,11 @@ pragma solidity ^0.8.0;
 contract ReceiverContract {
     string public lastMessage;
     address public lastSender;
-    uint16 public lastSrcChainId;
+    uint256 public lastSrcChainId;
     
-    event MessageReceived(string message, address sender, uint16 srcChainId);
+    event MessageReceived(string message, address sender, uint256 srcChainId);
 
-    function lzReceive(uint16 srcChainId, bytes calldata srcAddress, bytes calldata payload) external {
+    function lzReceive(uint256 srcChainId, bytes calldata srcAddress, bytes calldata payload) external {
         string memory msgDecoded = abi.decode(payload, (string));
         address senderAddress = abi.decode(srcAddress, (address));
         
@@ -20,7 +20,7 @@ contract ReceiverContract {
         emit MessageReceived(msgDecoded, senderAddress, srcChainId);
     }
     
-    function getLastMessageInfo() external view returns (string memory, address, uint16) {
+    function getLastMessageInfo() external view returns (string memory, address, uint256) {
         return (lastMessage, lastSender, lastSrcChainId);
     }
 }
